src/third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/custom/lifecycle-created-creation-api-expected.txt

CONSOLE ERROR: line 13: Uncaught TypeError: document.registerElement is not a function
This test ensures that the lifecycle callbacks of API-originated elements are visible in following script block.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

FAIL document.createElement('x-foo').callbacksCalled should be true (of type boolean). Was undefined (of type undefined).
FAIL document.createElement('div', 'x-bar').callbacksCalled should be true (of type boolean). Was undefined (of type undefined).
FAIL document.createElementNS('http://www.w3.org/1999/xhtml', 'x-foo').callbacksCalled should be true (of type boolean). Was undefined (of type undefined).
FAIL document.createElementNS('http://www.w3.org/1999/xhtml', 'div', 'x-bar').callbacksCalled should be true (of type boolean). Was undefined (of type undefined).
FAIL foo.cloneNode().callbacksCalled should be true (of type boolean). Was undefined (of type undefined).
FAIL foo.cloneNode(true).firstChild.callbacksCalled should be true (of type boolean). Was undefined (of type undefined).
FAIL bar.cloneNode().callbacksCalled should be true (of type boolean). Was undefined (of type undefined).
FAIL bar.cloneNode(true).firstChild.callbacksCalled should be true (of type boolean). Was undefined (of type undefined).
FAIL foreignFoo.callbacksCalled should be true (of type boolean). Was undefined (of type undefined).
FAIL foreignFoo.firstChild.callbacksCalled should be true (of type boolean). Was undefined (of type undefined).
FAIL importedFoo.callbacksCalled should be true (of type boolean). Was undefined (of type undefined).
FAIL importedFoo.firstChild.callbacksCalled should be true (of type boolean). Was undefined (of type undefined).
FAIL window.callbacksCalled should be div,x-foo. Was .
FAIL importedBar.callbacksCalled should be true (of type boolean). Was undefined (of type undefined).
FAIL importedBar.firstChild.callbacksCalled should be true (of type boolean). Was undefined (of type undefined).
FAIL window.callbacksCalled should be DIV,X-FOO. Was .
FAIL window.callbacksCalled should be X-FOO. Was .
FAIL window.callbacksCalled should be X-FOO. Was .
FAIL successfullyParsed should be true. Was false.

TEST COMPLETE
